Fix asset state when changing users
