+++ client/src/core/systems/models.ts
@@ -1,1 +1,1 @@
-export type SystemClearReason = "full-loading" | "partial-loading" | "leaving";
+export type SystemClearReason = "full-loading" | "partial-loading" | "leaving" | "logging-out";
